Introduce Helm Chart Test Framework Along With Basic Tests (#9)
We had no tests of the Helm charts in the past. This fixes that.

- Fix a few issues that got revealed by the tests.
- Introduce new CI steps
- Introduce linters

The tests are based on `gruntwork-io/terratest` which allow you to
assert certain facts about the Kubernetes environment. The basic set of
tests installs a chart using the minimal set of variables and asserts
that the pods start succesfully.

We now test MLS validation, payer service and the main sync+API service.
I will make a follow-up PR that splits the sync service from the API
service.

The majority of the code here is required to handle the weird Github
actions environment that does not allow for easy log collection. The
framework collects all relevant logs, dumps them to a file and collects
all these files as a CI artefact.

Each test runs in its own namespace, so in theory, we could run them all
in parallel. This also allows for easy cleanup for local testing as all
you have to do is delete the namespace.

We can use this framework for much more complicated stuff if we felt
like it. For now, it just ensures that the charts actually work.

Fixes #1